TemplateFinder: fix Coverity ID 1229295 Unintentional integer overflow
The expression's value may not be what the programmer intended, because the
expression is evaluated using a narrow (i.e. few bits) integer type.

In TemplateFinder::TemplateFinder(PGMConverter *, BorderDetector *,
EdgeDetector *, MythPlayer *, int, QString): An integer overflow occurs,
with the result converted to a wider integer type.

+++ mythtv/programs/mythcommflag/TemplateFinder.cpp
@@ -746,7 +746,7 @@ TemplateFinder::TemplateFinder(PGMConverter *pgmc, BorderDetector *bd,
     const float fps = player->GetFrameRate();
 
     frameInterval = (int)roundf(sampleTime * fps / samplesNeeded);
-    endFrame = 0 + frameInterval * samplesNeeded - 1;
+    endFrame = 0 + (long long)frameInterval * samplesNeeded - 1;
 
     LOG(VB_COMMFLAG, LOG_INFO,
         QString("TemplateFinder: sampleTime=%1s, samplesNeeded=%2, endFrame=%3")
